clean up navigation, remove unused components
