+++ components/header.tsx
@@ -1,69 +1,40 @@
 import { Box, Tabs } from '@bigcommerce/big-design';
 import { useRouter } from 'next/router';
 import { useEffect, useState } from 'react';
-import InnerHeader from './innerHeader';
 
 export const TabIds = {
-    HOME: 'home',
-    PRODUCTS: 'products',
     SEGMENTS: 'segments',
     CUSTOMERS: 'customers',
 };
 
 export const TabRoutes = {
-    [TabIds.HOME]: '/',
-    [TabIds.PRODUCTS]: '/products',
-    [TabIds.SEGMENTS]: '/segments',
+    [TabIds.SEGMENTS]: '/',
     [TabIds.CUSTOMERS]: '/customers',
 };
 
-const HeaderlessRoutes = [
-    '/orders/[orderId]/labels',
-    '/orders/[orderId]/modal',
-]
-
-const InnerRoutes = [
-    '/products/[pid]',
-];
-
 const HeaderTypes = {
-    GLOBAL: 'global',
-    INNER: 'inner',
-    HEADERLESS: 'headerless',
+    GLOBAL: 'global'
 };
 
 const Header = () => {
     const [activeTab, setActiveTab] = useState<string>('');
-    const [headerType, setHeaderType] = useState<string>(HeaderTypes.GLOBAL);
     const router = useRouter();
     const { pathname } = router;
 
     useEffect(() => {
-        if (InnerRoutes.includes(pathname)) {
-            // Use InnerHeader if route matches inner routes
-            setHeaderType(HeaderTypes.INNER);
-        } else if (HeaderlessRoutes.includes(pathname)) {
-            setHeaderType(HeaderTypes.HEADERLESS);
-        } else {
-            // Check if new route matches TabRoutes
-            const tabKey = Object.keys(TabRoutes).find(key => TabRoutes[key] === pathname);
+        // Check if new route matches TabRoutes
+        const tabKey = Object.keys(TabRoutes).find(key => TabRoutes[key] === pathname);
 
-            // Set the active tab to tabKey or set no active tab if route doesn't match (404)
-            setActiveTab(tabKey ?? '');
-            setHeaderType(HeaderTypes.GLOBAL);
-        }
-
+        // Set the active tab to tabKey or set no active tab if route doesn't match (404)
+        setActiveTab(tabKey ?? '');
     }, [pathname]);
 
     useEffect(() => {
-        // Prefetch products page to reduce latency (doesn't prefetch in dev)
-        router.prefetch('/products');
-        router.prefetch('/segments');
+        // Prefetch segments page to reduce latency (doesn't prefetch in dev)
+        router.prefetch('/');
     });
 
     const items = [
-        { id: TabIds.HOME, title: 'Home' },
-        { id: TabIds.PRODUCTS, title: 'Products' },
         { id: TabIds.SEGMENTS, title: 'Segments' },
         { id: TabIds.CUSTOMERS, title: 'Customers' },
     ];
@@ -73,9 +44,6 @@ const Header = () => {
 
         return router.push(TabRoutes[tabId]);
     };
-
-    if (headerType === HeaderTypes.HEADERLESS) return null;
-    if (headerType === HeaderTypes.INNER) return <InnerHeader />;
 
     return (
         <Box marginBottom="xxLarge">
