+++ pages/_app.tsx
@@ -1,4 +1,4 @@
-import { Box, GlobalStyles } from '@bigcommerce/big-design';
+import { Box, GlobalStyles, H1 } from '@bigcommerce/big-design';
 import { theme as defaultTheme } from '@bigcommerce/big-design-theme';
 import type { AppProps } from 'next/app';
 import { ThemeProvider } from 'styled-components';
@@ -21,6 +21,7 @@ const MyApp = ({ Component, pageProps }: AppProps) => {
                 marginHorizontal={{ mobile: 'none', tablet: 'xxxLarge' }}
                 marginVertical={{ mobile: 'none', tablet: "xxLarge" }}
             >
+                <H1>Customer Segment Manager</H1>
                 <Header />
                 <SessionProvider>
                     <Component {...pageProps} />
